edits to webpages A and B
